chore(refactor): Move `EvalError` and related structs into `oneil_output`
