+++ src-rs/oneil_eval/src/lib.rs
@@ -2,15 +2,12 @@
 //! Evaluator for the Oneil programming language
 
 mod context;
-pub mod error;
 mod eval_expr;
 mod eval_model;
 mod eval_parameter;
 mod eval_unit;
 
 pub use context::{ExternalEvaluationContext, IrLoadError};
-pub use error::{EvalError, EvalErrors};
-
 pub use eval_expr::eval_expr_in_model;
 pub use eval_model::eval_model_from_graph;
