+++ src-rs/oneil_output/src/error/model_eval_errors.rs
@@ -1,0 +1,18 @@
+use indexmap::{IndexMap, IndexSet};
+use oneil_shared::{
+    EvalInstanceKey,
+    symbols::{ParameterName, TestIndex},
+};
+
+use super::EvalError;
+
+/// Errors collected while evaluating a single model (parameters, tests, and broken references).
+#[derive(Debug, Clone)]
+pub struct ModelEvalErrors {
+    /// Errors that occurred during evaluation of the parameters.
+    pub parameters: IndexMap<ParameterName, Vec<EvalError>>,
+    /// Errors that occurred during evaluation of the tests.
+    pub tests: IndexMap<TestIndex, Vec<EvalError>>,
+    /// References that had errors.
+    pub references: IndexSet<EvalInstanceKey>,
+}
