+++ src-rs/oneil_runtime/src/output/mod.rs
@@ -7,8 +7,8 @@ pub use oneil_analysis::output::Independents;
 pub use oneil_ast as ast;
 pub use oneil_ir as ir;
 pub use oneil_output::{
-    BuiltinDependency, DebugInfo, DependencySet, ExternalDependency, Model, Number, Parameter,
-    ParameterDependency, PrintLevel, Test, TestResult, Unit, Value,
+    BuiltinDependency, DebugInfo, DependencySet, ExternalDependency, Model, ModelEvalErrors,
+    Number, Parameter, ParameterDependency, PrintLevel, Test, TestResult, Unit, Value,
 };
 pub use oneil_shared::{error::OneilDiagnostic, span::Span};
